Refactor file structure Relates #30
