+++ client/src/components/common/index.js
@@ -1,0 +1,25 @@
+import Avatar from './Avatar';
+import BackButton from './BackButton';
+import Button from './Button';
+import Checkbox from './Checkbox';
+import Dropdown from './Dropdown';
+import Input from './Input';
+import MessageCard from './MessageCard';
+import PageNotFound from './PageNotFound';
+import UnderConstruction from './UnderConstruction';
+import UserCard from './UserCard';
+import UserMenu from './UserMenu';
+
+export {
+  Avatar,
+  BackButton,
+  Button,
+  Checkbox,
+  Dropdown,
+  Input,
+  MessageCard,
+  PageNotFound,
+  UnderConstruction,
+  UserCard,
+  UserMenu,
+};
