+++ client/src/components/pages/Landing/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Button from '../../common/Button';
+import { Button } from '../../common';
 import { logo } from '../../assets/svgPaths';
 import './index.css';
