+++ client/src/components/pages/Settings/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
+import { UnderConstruction } from '../../common';
 import './index.css';
-import UnderConstruction from '../../common/UnderConstruction';
 
 const Settings = () => <UnderConstruction />;
