+++ client/src/components/pages/index.js
@@ -1,0 +1,9 @@
+import Login from './Login';
+import Signup from './Signup';
+import Profile from './Profile';
+import Landing from './Landing';
+import Home from './Home';
+import Chat from './Chat';
+import Settings from './Settings';
+
+export { Login, Signup, Profile, Landing, Home, Chat, Settings };
